broke out into components and added News
